backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-socketio>=5.11.0
google-generativeai>=0.3.0
requests>=2.31.0
httpx>=0.25.0
